Add dualstack network connections

+++ types.h
@@ -3,13 +3,15 @@
 #ifndef TYPES_H
 #define TYPES_H
 
+#include <stdint.h>
+
 typedef struct {
 	char *cgi;
 	bool debug;
 	bool help;
 	char *address;
 	char *log;
-	unsigned port;
+	uint16_t port;
 	char *docroot;
 } sws_options;
